Huge overhaul, now does transitions, borders, displays a new icon, etc

+++ app/src/main/java/com/steve/theshortreport/service/response/Clouds.java
@@ -5,14 +5,14 @@ import com.google.gson.annotations.Expose;
 public class Clouds {
 
     @Expose
-    private Integer all;
+    private Double all;
 
     /**
      *
      * @return
      * The all
      */
-    public Integer getAll() {
+    public Double getAll() {
         return all;
     }
 
@@ -21,7 +21,7 @@ public class Clouds {
      * @param all
      * The all
      */
-    public void setAll(Integer all) {
+    public void setAll(Double all) {
         this.all = all;
     }
