+++ app/src/main/java/com/steve/theshortreport/service/response/Coord.java
@@ -5,16 +5,16 @@ import com.google.gson.annotations.Expose;
 public class Coord {
 
     @Expose
-    private Integer lon;
+    private Double lon;
     @Expose
-    private Integer lat;
+    private Double lat;
 
     /**
      *
      * @return
      * The lon
      */
-    public Integer getLon() {
+    public Double getLon() {
         return lon;
     }
 
@@ -23,7 +23,7 @@ public class Coord {
      * @param lon
      * The lon
      */
-    public void setLon(Integer lon) {
+    public void setLon(Double lon) {
         this.lon = lon;
     }
 
@@ -32,7 +32,7 @@ public class Coord {
      * @return
      * The lat
      */
-    public Integer getLat() {
+    public Double getLat() {
         return lat;
     }
 
@@ -41,7 +41,7 @@ public class Coord {
      * @param lat
      * The lat
      */
-    public void setLat(Integer lat) {
+    public void setLat(Double lat) {
         this.lat = lat;
     }
